edited dune exploration

+++ task1/src/the_dune.kt
@@ -1,8 +1,8 @@
 fun main(){
-    println(exploreDune(height=20,climbRate=3))
+    println(exploreDune(height=20.0,climbRate=3.0))
 }
 
-fun exploreDune(height: Int, climbRate: Int): String {
-    var ClimbTime = (height.toDouble()/climbRate.toDouble())
+fun exploreDune(height: Double, climbRate: Double): String {
+    var ClimbTime = (height/climbRate)
     return "Time to Climb the Dune: %.2f".format(ClimbTime) + " hours"
 }
